requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
Django==3.2.5
djangorestframework==3.12.4
olefile==0.46
Pillow==8.3.1
pip==21.2.4
psycopg2==2.8.6
pycparser==2.20
pytz==2021.1
setuptools==52.0.0.post20210125
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
wheel==0.37.0
